+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## v0.0.65 (04/28/2020)
+## v0.0.66 (04/08/202)
+
+- Contributed the `Export Tour to Gist...` command to the `CodeTour` tree, which is available when you have the `CodeTour` extension installed.
+- Added a new `tour` gist type, that represent gists created by exporting a `CodeTour` as a gist. When you open tour gists, it will start playing the embedded tour.
+- User gists are now properly grouped after you follow a new user
+
+## v0.0.65 (04/06/2020)
 
 - When duplicating a file, you're now asked to specify the name of the new file before it's created. This prevents you from needing to duplicate and then rename the file, and therefore, enables the same thing in a single gesture.
 - Updated the "input tour" experience to use a custom-GistPad comment experience, as opposed to CodeTour (which isn't neccessary for a single comment affordance).